Moved live-stats to scorebooks/[id].
